inheritance and Polymorphism

+++ src/shapes/Square.java
@@ -1,16 +1,17 @@
 package shapes;
 
 public class Square extends Rectangle{
-    private int sides;
-    public Square (int side){
+    private double sides;
+    public Square (double side){
         super(side, side);
         sides = side;
     }
-    public int getArea(){
+    public double getArea(){
         System.out.println("This is a a square. Get the area!");
         return (int) Math.pow(sides,2);
+//        return length * length;
     }
-    public int getPerimeter(){
+    public double getPerimeter(){
         System.out.println("This is a square. Get the perimeter!");
         return sides * 4;
     }
